extract ImportantContacts grid from main DirectoryLanding component
