+++ app/(home)/Directory/index.tsx
@@ -1,37 +1,10 @@
 import * as React from 'react'
-import {Alert, StyleSheet, useWindowDimensions} from 'react-native'
+import {StyleSheet} from 'react-native'
 import {Stack, useLocalSearchParams, useRouter} from 'expo-router'
 import {useDispatch, useSelector} from 'react-redux'
 import {useQuery} from '@tanstack/react-query'
-import {
-	Button,
-	Grid,
-	HStack,
-	Host,
-	Image as UIImage,
-	List,
-	ProgressView,
-	Section,
-	Spacer,
-	Text as UIText,
-	VStack,
-} from '@expo/ui/swift-ui'
-import {
-	accessibilityElement,
-	accessibilityIdentifier,
-	accessibilityLabel,
-	buttonStyle,
-	font,
-	foregroundStyle,
-	frame,
-	listRowBackground,
-	listRowInsets,
-	listRowSeparator,
-	listStyle,
-	padding,
-	refreshable,
-} from '@expo/ui/swift-ui/modifiers'
-import * as c from '@frogpond/colors'
+import {Host, List, Section} from '@expo/ui/swift-ui'
+import {listStyle, refreshable} from '@expo/ui/swift-ui/modifiers'
 import {useDebounce} from '@frogpond/use-debounce'
 import {LoadingView, NoticeView} from '@frogpond/notice'
 import {DisclosureRow} from '../../../source/components/rows'
@@ -47,10 +20,8 @@ import {DirectoryResultsGrid} from '../../../source/features/directory/directory
 import {formatResults} from '../../../source/features/directory/helpers'
 import {directoryEntriesOptions} from '../../../source/features/directory/query'
 import {resolveSearch, searchHeading} from '../../../source/features/directory/resolve-search'
-import {ContactTile} from '../../../source/features/directory/tile'
-import {columnsForFontScale, inRows, TILE_SPACING} from '../../../source/components/tile-layout'
 import type {DirectoryItem, DirectorySearchTypeEnum} from '../../../source/features/directory/types'
-import {FILL_WIDTH, SCREEN_MARGIN} from '../../../source/features/home/button'
+import {ImportantContactsGrid} from '../../../source/features/directory/important-contacts-grid'
 
 function DirectoryView(): React.ReactNode {
 	let router = useRouter()
@@ -211,10 +182,6 @@ export default function DirectoryPage(): React.ReactNode {
 	)
 }
 
-/// Mirrored by TestIdentifiers.Directory.contactGrid.
-const CONTACT_GRID_ID = 'directory-contact-grid'
-const STALE_CONTACTS_LABEL = 'Contacts may be out of date'
-
 /**
  * What the Directory screen shows before a search: the curated campus contacts
  * as tiles, and the full campus department roster as an inset-grouped list.
@@ -225,34 +192,17 @@ const STALE_CONTACTS_LABEL = 'Contacts may be out of date'
  */
 function DirectoryLanding(): React.ReactNode {
 	let router = useRouter()
-	let {
-		data: contacts,
-		error: contactsError,
-		isLoading: contactsLoading,
-		refetch: refetchContacts,
-	} = useQuery(contactsOptions)
+	let contacts = useQuery(contactsOptions)
 	let {
 		data: departments,
 		isLoading: departmentsLoading,
 		refetch: refetchDepartments,
 	} = useQuery(directoryDepartmentsOptions)
-	let {fontScale} = useWindowDimensions()
-	let columns = columnsForFontScale(fontScale)
 
-	let showContactsError = React.useCallback(() => {
-		Alert.alert(
-			"Couldn't refresh contacts",
-			contactsError instanceof Error ? contactsError.message : 'Unknown error',
-			[
-				{text: 'Try Again', onPress: () => void refetchContacts()},
-				{text: 'OK', style: 'cancel'},
-			],
-		)
-	}, [contactsError, refetchContacts])
-
+	let refreshContacts = contacts.refetch
 	let refresh = React.useCallback(async () => {
-		await Promise.all([refetchContacts(), refetchDepartments()])
-	}, [refetchContacts, refetchDepartments])
+		await Promise.all([refreshContacts(), refetchDepartments()])
+	}, [refreshContacts, refetchDepartments])
 
 	return (
 		<Host matchContents={false} style={styles.host}>
@@ -264,78 +214,15 @@ function DirectoryLanding(): React.ReactNode {
 					}),
 				]}
 			>
-				<VStack
-					modifiers={[
-						listRowBackground('clear'),
-						listRowInsets({top: 0, leading: 0, bottom: 0, trailing: 0}),
-						listRowSeparator('hidden'),
-						// Horizontal + top only: the enclosing List owns the gap down
-						// to the Departments section below.
-						padding({horizontal: SCREEN_MARGIN, top: SCREEN_MARGIN}),
-						frame({maxWidth: FILL_WIDTH}),
-					]}
-					spacing={TILE_SPACING}
-				>
-					<HStack modifiers={[frame({maxWidth: FILL_WIDTH})]}>
-						<UIText modifiers={[font({textStyle: 'headline'})]}>Important Contacts</UIText>
-						<Spacer />
-						{contactsError && contacts ? (
-							<Button
-								modifiers={[buttonStyle('plain'), accessibilityLabel(STALE_CONTACTS_LABEL)]}
-								onPress={showContactsError}
-							>
-								<UIImage color={c.orange} systemName="exclamationmark.triangle.fill" />
-							</Button>
-						) : null}
-					</HStack>
-
-					{contacts ? (
-						<Grid
-							alignment="top"
-							horizontalSpacing={TILE_SPACING}
-							// The Grid itself carries no accessibility presence of its
-							// own, so accessibilityIdentifier alone lands on its first
-							// button descendant instead of the grid -- contain() gives it
-							// one, keeping the tiles as its individually-navigable
-							// children, which is what a UI test counting them needs.
-							modifiers={[
-								accessibilityElement('contain'),
-								accessibilityIdentifier(CONTACT_GRID_ID),
-							]}
-							verticalSpacing={TILE_SPACING}
-						>
-							{inRows(contacts, columns).map((row, i) => (
-								<Grid.Row key={i}>
-									{row.map((contact) => (
-										<ContactTile
-											key={contact.title}
-											contact={contact}
-											onPress={() =>
-												router.push({
-													pathname: '/Directory/named/[title]',
-													params: {title: contact.title},
-												})
-											}
-										/>
-									))}
-									{/* A short last row leaves its columns empty rather than
-									    stretching the tiles in it. 8 contacts divide evenly by
-									    4 and 2 columns but not by 3, so this padding matters at
-									    every column count, not just the edge cases. */}
-									{Array.from({length: columns - row.length}, (_, j) => (
-										<Spacer key={j} />
-									))}
-								</Grid.Row>
-							))}
-						</Grid>
-					) : contactsLoading ? (
-						<ProgressView />
-					) : (
-						<UIText modifiers={[foregroundStyle(c.secondaryLabel)]}>
-							Contacts are unavailable. Pull to try again.
-						</UIText>
-					)}
-				</VStack>
+				<ImportantContactsGrid
+					query={contacts}
+					onSelectContact={(contact) => {
+						router.push({
+							pathname: '/Directory/named/[title]',
+							params: {title: contact.title},
+						})
+					}}
+				/>
 
 				<DepartmentsList
 					departments={departments}
